refactor benchmark suite using criterion
