+++ Cargo.toml
@@ -20,7 +20,12 @@ serde = { version = "1.0", optional = true }
 [dev-dependencies]
 fnv = "1.0"         # Required bench tests.
 serde_json = "1.0"  # Required for testing the serde implementation.
-lazy_static = "1.0" # Required for bench tests.
+lazy_static = "1.0" # Required for benches.
+criterion = "0.3.1" # Required for benches.
+
+[[bench]]
+name = "bench"
+harness = false
 
 [features]
 default       = ["std", "serde-1"]
